requirements.txt
fastapi
uvicorn
supabase
python-dotenv
requests
beautifulsoup4
lxml
langchain-google-genai
gunicorn